GantryControl/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MQTT==1.0.7
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
paho-mqtt==1.5.1
python-dotenv==0.15.0
PyYAML==5.4.1
typing==3.7.4.3
Werkzeug==1.0.1
RPI.GPIO==0.7.0